+++ docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ### v0.3.0 (dev)
+- Isolated type identifier to another package
+- Added functions:
+  - Arithmetic: `ScalarExponent`, `ScalarModulo`
+- Modified existing functions to support new expression types
 ### v0.2.2 (Sep 25, 2021)
 - Added expressions: `Matrix`
 - Added functions: