+++ docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-### v0.3.0 (dev)
+### v0.3.1 (dev)
+- Added functions:
+  - Identify: `Type`
+### v0.3.0 (Nov 8, 2021)
 - Isolated type identifier to another package
 - Added expressions: `String`
 - Added functions: